+++ sdk_r23/frameworks/cinema.framework/source/c4d_tools.h
@@ -143,7 +143,7 @@ inline Float32 Modulo(Float32 a, Float32 b)
 		return 0.0;
 	Int32 n = (Int32) (a / b);
 
-	a -= n * b;
+	a -= ((Float32)n) * b;
 	if (a < 0.0)
 		a += b;
 
@@ -162,7 +162,7 @@ inline Float64 Modulo(Float64 a, Float64 b)
 		return 0.0;
 	Int32 n = (Int32) (a / b);
 
-	a -= n * b;
+	a -= ((Float64)n) * b;
 	if (a < 0.0)
 		a += b;
